requirements.txt
djangorestframework
